Refactor crops. Moved crop transforms to independent module and added type hints.
