+++ albumentations/augmentations/__init__.py
@@ -6,9 +6,13 @@ from .transforms import *
 
 # New transformations goes to individual files listed below
 from .domain_adaptation import *
+
 from .geometric.transforms import *
 from .geometric.functional import *
 from .geometric.resize import *
 from .geometric.rotate import *
+
+from .crops.transforms import *
+from .crops.functional import *
 
 from .utils import *
