+++ albumentations/augmentations/transforms.py
@@ -25,9 +25,7 @@ __all__ = [
     "Flip",
     "Normalize",
     "Transpose",
-    "RandomCrop",
     "RandomGamma",
-    "CenterCrop",
     "OpticalDistortion",
     "GridDistortion",
     "RandomGridShuffle",
@@ -52,13 +50,7 @@ __all__ = [
     "CoarseDropout",
     "ToFloat",
     "FromFloat",
-    "Crop",
-    "CropNonEmptyMaskIfExists",
-    "RandomSizedCrop",
-    "RandomResizedCrop",
     "RandomBrightnessContrast",
-    "RandomCropNearBBox",
-    "RandomSizedBBoxSafeCrop",
     "RandomSnow",
     "RandomRain",
     "RandomFog",
@@ -198,49 +190,6 @@ class PadIfNeeded(DualTransform):
         )
 
 
-class Crop(DualTransform):
-    """Crop region from image.
-
-    Args:
-        x_min (int): Minimum upper left x coordinate.
-        y_min (int): Minimum upper left y coordinate.
-        x_max (int): Maximum lower right x coordinate.
-        y_max (int): Maximum lower right y coordinate.
-
-    Targets:
-        image, mask, bboxes, keypoints
-
-    Image types:
-        uint8, float32
-    """
-
-    def __init__(self, x_min=0, y_min=0, x_max=1024, y_max=1024, always_apply=False, p=1.0):
-        super(Crop, self).__init__(always_apply, p)
-        self.x_min = x_min
-        self.y_min = y_min
-        self.x_max = x_max
-        self.y_max = y_max
-
-    def apply(self, img, **params):
-        return F.crop(img, x_min=self.x_min, y_min=self.y_min, x_max=self.x_max, y_max=self.y_max)
-
-    def apply_to_bbox(self, bbox, **params):
-        return F.bbox_crop(bbox, x_min=self.x_min, y_min=self.y_min, x_max=self.x_max, y_max=self.y_max, **params)
-
-    def apply_to_keypoint(self, keypoint, **params):
-        return F.crop_keypoint_by_coords(
-            keypoint,
-            crop_coords=(self.x_min, self.y_min, self.x_max, self.y_max),
-            crop_height=self.y_max - self.y_min,
-            crop_width=self.x_max - self.x_min,
-            rows=params["rows"],
-            cols=params["cols"],
-        )
-
-    def get_transform_init_args_names(self):
-        return ("x_min", "y_min", "x_max", "y_max")
-
-
 class VerticalFlip(DualTransform):
     """Flip the input vertically around the x-axis.
 
@@ -357,464 +306,6 @@ class Transpose(DualTransform):
 
     def get_transform_init_args_names(self):
         return ()
-
-
-class CenterCrop(DualTransform):
-    """Crop the central part of the input.
-
-    Args:
-        height (int): height of the crop.
-        width (int): width of the crop.
-        p (float): probability of applying the transform. Default: 1.
-
-    Targets:
-        image, mask, bboxes, keypoints
-
-    Image types:
-        uint8, float32
-
-    Note:
-        It is recommended to use uint8 images as input.
-        Otherwise the operation will require internal conversion
-        float32 -> uint8 -> float32 that causes worse performance.
-    """
-
-    def __init__(self, height, width, always_apply=False, p=1.0):
-        super(CenterCrop, self).__init__(always_apply, p)
-        self.height = height
-        self.width = width
-
-    def apply(self, img, **params):
-        return F.center_crop(img, self.height, self.width)
-
-    def apply_to_bbox(self, bbox, **params):
-        return F.bbox_center_crop(bbox, self.height, self.width, **params)
-
-    def apply_to_keypoint(self, keypoint, **params):
-        return F.keypoint_center_crop(keypoint, self.height, self.width, **params)
-
-    def get_transform_init_args_names(self):
-        return ("height", "width")
-
-
-class RandomCrop(DualTransform):
-    """Crop a random part of the input.
-
-    Args:
-        height (int): height of the crop.
-        width (int): width of the crop.
-        p (float): probability of applying the transform. Default: 1.
-
-    Targets:
-        image, mask, bboxes, keypoints
-
-    Image types:
-        uint8, float32
-    """
-
-    def __init__(self, height, width, always_apply=False, p=1.0):
-        super(RandomCrop, self).__init__(always_apply, p)
-        self.height = height
-        self.width = width
-
-    def apply(self, img, h_start=0, w_start=0, **params):
-        return F.random_crop(img, self.height, self.width, h_start, w_start)
-
-    def get_params(self):
-        return {"h_start": random.random(), "w_start": random.random()}
-
-    def apply_to_bbox(self, bbox, **params):
-        return F.bbox_random_crop(bbox, self.height, self.width, **params)
-
-    def apply_to_keypoint(self, keypoint, **params):
-        return F.keypoint_random_crop(keypoint, self.height, self.width, **params)
-
-    def get_transform_init_args_names(self):
-        return ("height", "width")
-
-
-class RandomCropNearBBox(DualTransform):
-    """Crop bbox from image with random shift by x,y coordinates
-
-    Args:
-        max_part_shift (float): float value in (0.0, 1.0) range. Default 0.3
-        p (float): probability of applying the transform. Default: 1.
-
-    Targets:
-        image, mask, bboxes, keypoints
-
-    Image types:
-        uint8, float32
-    """
-
-    def __init__(self, max_part_shift=0.3, always_apply=False, p=1.0):
-        super(RandomCropNearBBox, self).__init__(always_apply, p)
-        self.max_part_shift = max_part_shift
-
-    def apply(self, img, x_min=0, x_max=0, y_min=0, y_max=0, **params):
-        return F.clamping_crop(img, x_min, y_min, x_max, y_max)
-
-    def get_params_dependent_on_targets(self, params):
-        bbox = params["cropping_bbox"]
-        h_max_shift = int((bbox[3] - bbox[1]) * self.max_part_shift)
-        w_max_shift = int((bbox[2] - bbox[0]) * self.max_part_shift)
-
-        x_min = bbox[0] - random.randint(-w_max_shift, w_max_shift)
-        x_max = bbox[2] + random.randint(-w_max_shift, w_max_shift)
-
-        y_min = bbox[1] - random.randint(-h_max_shift, h_max_shift)
-        y_max = bbox[3] + random.randint(-h_max_shift, h_max_shift)
-
-        return {"x_min": x_min, "x_max": x_max, "y_min": y_min, "y_max": y_max}
-
-    def apply_to_bbox(self, bbox, x_min=0, x_max=0, y_min=0, y_max=0, **params):
-        h_start = y_min
-        w_start = x_min
-        return F.bbox_crop(bbox, y_max - y_min, x_max - x_min, h_start, w_start, **params)
-
-    def apply_to_keypoint(self, keypoint, x_min=0, x_max=0, y_min=0, y_max=0, **params):
-        return F.crop_keypoint_by_coords(
-            keypoint,
-            crop_coords=(x_min, y_min, x_max, y_max),
-            crop_height=y_max - y_min,
-            crop_width=x_max - x_min,
-            rows=params["rows"],
-            cols=params["cols"],
-        )
-
-    @property
-    def targets_as_params(self):
-        return ["cropping_bbox"]
-
-    def get_transform_init_args_names(self):
-        return ("max_part_shift",)
-
-
-class _BaseRandomSizedCrop(DualTransform):
-    # Base class for RandomSizedCrop and RandomResizedCrop
-
-    def __init__(self, height, width, interpolation=cv2.INTER_LINEAR, always_apply=False, p=1.0):
-        super(_BaseRandomSizedCrop, self).__init__(always_apply, p)
-        self.height = height
-        self.width = width
-        self.interpolation = interpolation
-
-    def apply(self, img, crop_height=0, crop_width=0, h_start=0, w_start=0, interpolation=cv2.INTER_LINEAR, **params):
-        crop = F.random_crop(img, crop_height, crop_width, h_start, w_start)
-        return FGeometric.resize(crop, self.height, self.width, interpolation)
-
-    def apply_to_bbox(self, bbox, crop_height=0, crop_width=0, h_start=0, w_start=0, rows=0, cols=0, **params):
-        return F.bbox_random_crop(bbox, crop_height, crop_width, h_start, w_start, rows, cols)
-
-    def apply_to_keypoint(self, keypoint, crop_height=0, crop_width=0, h_start=0, w_start=0, rows=0, cols=0, **params):
-        keypoint = F.keypoint_random_crop(keypoint, crop_height, crop_width, h_start, w_start, rows, cols)
-        scale_x = self.width / crop_width
-        scale_y = self.height / crop_height
-        keypoint = FGeometric.keypoint_scale(keypoint, scale_x, scale_y)
-        return keypoint
-
-
-class RandomSizedCrop(_BaseRandomSizedCrop):
-    """Crop a random part of the input and rescale it to some size.
-
-    Args:
-        min_max_height ((int, int)): crop size limits.
-        height (int): height after crop and resize.
-        width (int): width after crop and resize.
-        w2h_ratio (float): aspect ratio of crop.
-        interpolation (OpenCV flag): flag that is used to specify the interpolation algorithm. Should be one of:
-            cv2.INTER_NEAREST, cv2.INTER_LINEAR, cv2.INTER_CUBIC, cv2.INTER_AREA, cv2.INTER_LANCZOS4.
-            Default: cv2.INTER_LINEAR.
-        p (float): probability of applying the transform. Default: 1.
-
-    Targets:
-        image, mask, bboxes, keypoints
-
-    Image types:
-        uint8, float32
-    """
-
-    def __init__(
-        self, min_max_height, height, width, w2h_ratio=1.0, interpolation=cv2.INTER_LINEAR, always_apply=False, p=1.0
-    ):
-        super(RandomSizedCrop, self).__init__(
-            height=height, width=width, interpolation=interpolation, always_apply=always_apply, p=p
-        )
-        self.min_max_height = min_max_height
-        self.w2h_ratio = w2h_ratio
-
-    def get_params(self):
-        crop_height = random.randint(self.min_max_height[0], self.min_max_height[1])
-        return {
-            "h_start": random.random(),
-            "w_start": random.random(),
-            "crop_height": crop_height,
-            "crop_width": int(crop_height * self.w2h_ratio),
-        }
-
-    def get_transform_init_args_names(self):
-        return "min_max_height", "height", "width", "w2h_ratio", "interpolation"
-
-
-class RandomResizedCrop(_BaseRandomSizedCrop):
-    """Torchvision's variant of crop a random part of the input and rescale it to some size.
-
-    Args:
-        height (int): height after crop and resize.
-        width (int): width after crop and resize.
-        scale ((float, float)): range of size of the origin size cropped
-        ratio ((float, float)): range of aspect ratio of the origin aspect ratio cropped
-        interpolation (OpenCV flag): flag that is used to specify the interpolation algorithm. Should be one of:
-            cv2.INTER_NEAREST, cv2.INTER_LINEAR, cv2.INTER_CUBIC, cv2.INTER_AREA, cv2.INTER_LANCZOS4.
-            Default: cv2.INTER_LINEAR.
-        p (float): probability of applying the transform. Default: 1.
-
-    Targets:
-        image, mask, bboxes, keypoints
-
-    Image types:
-        uint8, float32
-    """
-
-    def __init__(
-        self,
-        height,
-        width,
-        scale=(0.08, 1.0),
-        ratio=(0.75, 1.3333333333333333),
-        interpolation=cv2.INTER_LINEAR,
-        always_apply=False,
-        p=1.0,
-    ):
-
-        super(RandomResizedCrop, self).__init__(
-            height=height, width=width, interpolation=interpolation, always_apply=always_apply, p=p
-        )
-        self.scale = scale
-        self.ratio = ratio
-
-    def get_params_dependent_on_targets(self, params):
-        img = params["image"]
-        area = img.shape[0] * img.shape[1]
-
-        for _attempt in range(10):
-            target_area = random.uniform(*self.scale) * area
-            log_ratio = (math.log(self.ratio[0]), math.log(self.ratio[1]))
-            aspect_ratio = math.exp(random.uniform(*log_ratio))
-
-            w = int(round(math.sqrt(target_area * aspect_ratio)))  # skipcq: PTC-W0028
-            h = int(round(math.sqrt(target_area / aspect_ratio)))  # skipcq: PTC-W0028
-
-            if 0 < w <= img.shape[1] and 0 < h <= img.shape[0]:
-                i = random.randint(0, img.shape[0] - h)
-                j = random.randint(0, img.shape[1] - w)
-                return {
-                    "crop_height": h,
-                    "crop_width": w,
-                    "h_start": i * 1.0 / (img.shape[0] - h + 1e-10),
-                    "w_start": j * 1.0 / (img.shape[1] - w + 1e-10),
-                }
-
-        # Fallback to central crop
-        in_ratio = img.shape[1] / img.shape[0]
-        if in_ratio < min(self.ratio):
-            w = img.shape[1]
-            h = int(round(w / min(self.ratio)))
-        elif in_ratio > max(self.ratio):
-            h = img.shape[0]
-            w = int(round(h * max(self.ratio)))
-        else:  # whole image
-            w = img.shape[1]
-            h = img.shape[0]
-        i = (img.shape[0] - h) // 2
-        j = (img.shape[1] - w) // 2
-        return {
-            "crop_height": h,
-            "crop_width": w,
-            "h_start": i * 1.0 / (img.shape[0] - h + 1e-10),
-            "w_start": j * 1.0 / (img.shape[1] - w + 1e-10),
-        }
-
-    def get_params(self):
-        return {}
-
-    @property
-    def targets_as_params(self):
-        return ["image"]
-
-    def get_transform_init_args_names(self):
-        return "height", "width", "scale", "ratio", "interpolation"
-
-
-class RandomSizedBBoxSafeCrop(DualTransform):
-    """Crop a random part of the input and rescale it to some size without loss of bboxes.
-
-    Args:
-        height (int): height after crop and resize.
-        width (int): width after crop and resize.
-        erosion_rate (float): erosion rate applied on input image height before crop.
-        interpolation (OpenCV flag): flag that is used to specify the interpolation algorithm. Should be one of:
-            cv2.INTER_NEAREST, cv2.INTER_LINEAR, cv2.INTER_CUBIC, cv2.INTER_AREA, cv2.INTER_LANCZOS4.
-            Default: cv2.INTER_LINEAR.
-        p (float): probability of applying the transform. Default: 1.
-
-    Targets:
-        image, mask, bboxes
-
-    Image types:
-        uint8, float32
-    """
-
-    def __init__(self, height, width, erosion_rate=0.0, interpolation=cv2.INTER_LINEAR, always_apply=False, p=1.0):
-        super(RandomSizedBBoxSafeCrop, self).__init__(always_apply, p)
-        self.height = height
-        self.width = width
-        self.interpolation = interpolation
-        self.erosion_rate = erosion_rate
-
-    def apply(self, img, crop_height=0, crop_width=0, h_start=0, w_start=0, interpolation=cv2.INTER_LINEAR, **params):
-        crop = F.random_crop(img, crop_height, crop_width, h_start, w_start)
-        return FGeometric.resize(crop, self.height, self.width, interpolation)
-
-    def get_params_dependent_on_targets(self, params):
-        img_h, img_w = params["image"].shape[:2]
-        if len(params["bboxes"]) == 0:  # less likely, this class is for use with bboxes.
-            erosive_h = int(img_h * (1.0 - self.erosion_rate))
-            crop_height = img_h if erosive_h >= img_h else random.randint(erosive_h, img_h)
-            return {
-                "h_start": random.random(),
-                "w_start": random.random(),
-                "crop_height": crop_height,
-                "crop_width": int(crop_height * img_w / img_h),
-            }
-        # get union of all bboxes
-        x, y, x2, y2 = union_of_bboxes(
-            width=img_w, height=img_h, bboxes=params["bboxes"], erosion_rate=self.erosion_rate
-        )
-        # find bigger region
-        bx, by = x * random.random(), y * random.random()
-        bx2, by2 = x2 + (1 - x2) * random.random(), y2 + (1 - y2) * random.random()
-        bw, bh = bx2 - bx, by2 - by
-        crop_height = img_h if bh >= 1.0 else int(img_h * bh)
-        crop_width = img_w if bw >= 1.0 else int(img_w * bw)
-        h_start = np.clip(0.0 if bh >= 1.0 else by / (1.0 - bh), 0.0, 1.0)
-        w_start = np.clip(0.0 if bw >= 1.0 else bx / (1.0 - bw), 0.0, 1.0)
-        return {"h_start": h_start, "w_start": w_start, "crop_height": crop_height, "crop_width": crop_width}
-
-    def apply_to_bbox(self, bbox, crop_height=0, crop_width=0, h_start=0, w_start=0, rows=0, cols=0, **params):
-        return F.bbox_random_crop(bbox, crop_height, crop_width, h_start, w_start, rows, cols)
-
-    @property
-    def targets_as_params(self):
-        return ["image", "bboxes"]
-
-    def get_transform_init_args_names(self):
-        return ("height", "width", "erosion_rate", "interpolation")
-
-
-class CropNonEmptyMaskIfExists(DualTransform):
-    """Crop area with mask if mask is non-empty, else make random crop.
-
-    Args:
-        height (int): vertical size of crop in pixels
-        width (int): horizontal size of crop in pixels
-        ignore_values (list of int): values to ignore in mask, `0` values are always ignored
-            (e.g. if background value is 5 set `ignore_values=[5]` to ignore)
-        ignore_channels (list of int): channels to ignore in mask
-            (e.g. if background is a first channel set `ignore_channels=[0]` to ignore)
-        p (float): probability of applying the transform. Default: 1.0.
-
-    Targets:
-        image, mask, bboxes, keypoints
-
-    Image types:
-        uint8, float32
-    """
-
-    def __init__(self, height, width, ignore_values=None, ignore_channels=None, always_apply=False, p=1.0):
-        super(CropNonEmptyMaskIfExists, self).__init__(always_apply, p)
-
-        if ignore_values is not None and not isinstance(ignore_values, list):
-            raise ValueError("Expected `ignore_values` of type `list`, got `{}`".format(type(ignore_values)))
-        if ignore_channels is not None and not isinstance(ignore_channels, list):
-            raise ValueError("Expected `ignore_channels` of type `list`, got `{}`".format(type(ignore_channels)))
-
-        self.height = height
-        self.width = width
-        self.ignore_values = ignore_values
-        self.ignore_channels = ignore_channels
-
-    def apply(self, img, x_min=0, x_max=0, y_min=0, y_max=0, **params):
-        return F.crop(img, x_min, y_min, x_max, y_max)
-
-    def apply_to_bbox(self, bbox, x_min=0, x_max=0, y_min=0, y_max=0, **params):
-        return F.bbox_crop(
-            bbox, x_min=x_min, x_max=x_max, y_min=y_min, y_max=y_max, rows=params["rows"], cols=params["cols"]
-        )
-
-    def apply_to_keypoint(self, keypoint, x_min=0, x_max=0, y_min=0, y_max=0, **params):
-        return F.crop_keypoint_by_coords(
-            keypoint,
-            crop_coords=[x_min, y_min, x_max, y_max],
-            crop_height=y_max - y_min,
-            crop_width=x_max - x_min,
-            rows=params["rows"],
-            cols=params["cols"],
-        )
-
-    def _preprocess_mask(self, mask):
-        mask_height, mask_width = mask.shape[:2]
-
-        if self.ignore_values is not None:
-            ignore_values_np = np.array(self.ignore_values)
-            mask = np.where(np.isin(mask, ignore_values_np), 0, mask)
-
-        if mask.ndim == 3 and self.ignore_channels is not None:
-            target_channels = np.array([ch for ch in range(mask.shape[-1]) if ch not in self.ignore_channels])
-            mask = np.take(mask, target_channels, axis=-1)
-
-        if self.height > mask_height or self.width > mask_width:
-            raise ValueError(
-                "Crop size ({},{}) is larger than image ({},{})".format(
-                    self.height, self.width, mask_height, mask_width
-                )
-            )
-
-        return mask
-
-    def update_params(self, params, **kwargs):
-        if "mask" in kwargs:
-            mask = self._preprocess_mask(kwargs["mask"])
-        elif "masks" in kwargs and len(kwargs["masks"]):
-            masks = kwargs["masks"]
-            mask = self._preprocess_mask(masks[0])
-            for m in masks[1:]:
-                mask |= self._preprocess_mask(m)
-        else:
-            raise RuntimeError("Can not find mask for CropNonEmptyMaskIfExists")
-
-        mask_height, mask_width = mask.shape[:2]
-
-        if mask.any():
-            mask = mask.sum(axis=-1) if mask.ndim == 3 else mask
-            non_zero_yx = np.argwhere(mask)
-            y, x = random.choice(non_zero_yx)
-            x_min = x - random.randint(0, self.width - 1)
-            y_min = y - random.randint(0, self.height - 1)
-            x_min = np.clip(x_min, 0, mask_width - self.width)
-            y_min = np.clip(y_min, 0, mask_height - self.height)
-        else:
-            x_min = random.randint(0, mask_width - self.width)
-            y_min = random.randint(0, mask_height - self.height)
-
-        x_max = x_min + self.width
-        y_max = y_min + self.height
-
-        params.update({"x_min": x_min, "x_max": x_max, "y_min": y_min, "y_max": y_max})
-        return params
-
-    def get_transform_init_args_names(self):
-        return ("height", "width", "ignore_values", "ignore_channels")
 
 
 class OpticalDistortion(DualTransform):
